backend/requirements.txt
anchorpy==0.21.0
anchorpy-core==0.2.0
anyio==4.10.0
asgiref==3.9.1
async-generator==1.10
attrs==25.3.0
base58==2.1.1
based58==0.1.1
beautifulsoup4==4.14.2
borsh-construct==0.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
construct==2.10.68
construct-typing==0.5.6
cryptography==46.0.1
Django==5.2.6
django-allauth==65.1.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipfshttpclient==0.8.0a2
jsonalias==0.1.1
lxml==6.0.2
morphys==1.0
multiaddr==0.0.9
mypy_extensions==1.1.0
netaddr==1.3.0
outcome==1.3.0.post0
packaging==25.0
pillow==10.4.0
pluggy==1.6.0
psutil==7.1.0
psycopg2-binary==2.9.11
py-cid==0.3.1
py-multibase==1.0.3
py-multicodec==0.2.1
pycparser==2.23
Pygments==2.19.2
pyheck==0.1.5
PyJWT==2.9.0
pymultihash==0.8.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-baseconv==1.2.2
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
stripe==13.1.0
sumtypes==0.1a6
toml==0.10.2
toolz==0.11.2
trio==0.31.0
trio-typing==0.10.0
typing_extensions==4.15.0
urllib3==2.5.0
varint==1.0.2
websockets==15.0
whitenoise==6.11.0
zipp==3.23.0
